Splitting Iroha chain listener (#377)
