+++ notifications/src/main/kotlin/com/d3/notifications/config/NotificationAppConfiguration.kt
@@ -1,5 +1,8 @@
 package com.d3.notifications.config
 
+import com.d3.notifications.provider.D3ClientProvider
+import com.d3.notifications.push.PushServiceFactory
+import com.d3.notifications.smtp.SMTPServiceImpl
 import config.getConfigFolder
 import config.loadRawConfigs
 import io.grpc.ManagedChannelBuilder
@@ -7,9 +10,6 @@ import jp.co.soramitsu.iroha.java.IrohaAPI
 import jp.co.soramitsu.iroha.java.QueryAPI
 import model.IrohaCredential
 import nl.martijndwars.webpush.PushService
-import com.d3.notifications.provider.D3ClientProvider
-import com.d3.notifications.push.PushServiceFactory
-import com.d3.notifications.smtp.SMTPServiceImpl
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import sidechain.iroha.IrohaChainListener
